// extensions/shortestpath.oj/src/generated/api-contract.ts
export type LocalJudging = {
  checker?: LocalJudgingChecker;
  files: Array<LocalJudgingFile>;
  grader?: LocalJudgingGrader;
  kind: string;
  schema_version: number;
};

export type LocalJudgingChecker = {
  answer_kind: string;
  arguments: Array<string>;
};

export type LocalJudgingFile = {
  name: string;
  role: string;
  url: string;
};

export type LocalJudgingGrader = {
  entry_point: string;
  example?: LocalJudgingGraderExample;
  input: LocalJudgingGraderInput;
  output: string;
};

export type LocalJudgingGraderExample = {
  input: string;
  output?: string;
};

export type LocalJudgingGraderInput = {
  before?: string;
  format: string;
  function?: string;
  mode: string;
};